Added front end UI

+++ backend/controllers/goalController.js
@@ -1,13 +1,15 @@
 const express = require('express')
 const asyncHandler = require('express-async-handler')
-
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
+
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id})
     
     res.status(200).json(goals)
 })
@@ -24,6 +26,7 @@ const setGoal = asyncHandler(async(req, res) => {
     
     const goal = await Goal.create({
         text:req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -39,6 +42,18 @@ const updateGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
     
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new : true})
@@ -58,7 +73,18 @@ const deleteGoal =asyncHandler( async(req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
-    
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await goal.deleteOne()
 
     res.status(200).json({ id: req.params.id })
